[small] refactor import/export method for components + add models folder with user and suggested feed
