+++ components/index.ts
@@ -1,0 +1,6 @@
+export * from "./ExternalLink";
+export * from "./Footer";
+export * from "./Header";
+export * from "./HeadMeta";
+export * from "./NewFeedForm";
+export * from "./ProgressLoader";
