+++ models/SuggestedFeed.ts
@@ -1,0 +1,5 @@
+export interface SuggestedFeed {
+    SiteUrl: string;
+    FeedUrl: string;
+    Description: string;
+}
