+++ models/UserFeed.ts
@@ -1,0 +1,4 @@
+export interface UserFeed {
+    SiteUrl: string;
+    VisitedPaths: string[];
+}
