+++ models/index.ts
@@ -1,0 +1,2 @@
+export * from "./UserFeed";
+export * from "./SuggestedFeed";
